+++ raspberry/server/services/state.service.js
@@ -62,6 +62,7 @@ class StateService {
       hdmi0: false,
       hdmi1: false,
       wrongPort: false,
+      hdmiSwapped: false,
       updatedAt: null,
     };